+++ src/detection/brightness/brightness_nosupport.c
@@ -1,0 +1,6 @@
+#include "brightness.h"
+
+const char* ffDetectBrightness(FF_UNUSED_PARAM FFlist* result)
+{
+    return "Not supported on this platform";
+}